Decouple the code
