+++ src/fheroes2/battle/battle_interface.cpp
@@ -38,8 +38,8 @@
 #include "battle_tower.h"
 #include "battle_troop.h"
 #include "castle.h"
-#include "game.h"
 #include "game_delays.h"
+#include "game_hotkeys.h"
 #include "ground.h"
 #include "icn.h"
 #include "interface_list.h"
