+++ src/fheroes2/castle/castle_mageguild.cpp
@@ -28,7 +28,7 @@
 #include "castle.h"
 #include "cursor.h"
 #include "dialog.h"
-#include "game.h"
+#include "game_hotkeys.h"
 #include "icn.h"
 #include "mageguild.h"
 #include "race.h"
