+++ src/fheroes2/castle/castle_town.cpp
@@ -30,7 +30,7 @@
 #include "castle_ui.h"
 #include "cursor.h"
 #include "dialog.h"
-#include "game.h"
+#include "game_hotkeys.h"
 #include "heroes.h"
 #include "icn.h"
 #include "kingdom.h"
