+++ src/fheroes2/castle/castle_well.cpp
@@ -29,8 +29,8 @@
 #include "castle.h"
 #include "cursor.h"
 #include "dialog.h"
-#include "game.h"
 #include "game_delays.h"
+#include "game_hotkeys.h"
 #include "icn.h"
 #include "kingdom.h"
 #include "monster_anim.h"
