+++ src/fheroes2/dialog/dialog_armyinfo.cpp
@@ -29,8 +29,8 @@
 #include "battle_cell.h"
 #include "cursor.h"
 #include "dialog.h"
-#include "game.h"
 #include "game_delays.h"
+#include "game_hotkeys.h"
 #include "icn.h"
 #include "luck.h"
 #include "monster.h"
