+++ src/fheroes2/dialog/dialog_chest.cpp
@@ -24,7 +24,7 @@
 #include "agg_image.h"
 #include "cursor.h"
 #include "dialog.h"
-#include "game.h"
+#include "game_hotkeys.h"
 #include "icn.h"
 #include "settings.h"
 #include "text.h"
