+++ src/fheroes2/dialog/dialog_system_options.cpp
@@ -25,6 +25,7 @@
 #include "dialog.h"
 #include "game.h"
 #include "game_delays.h"
+#include "game_hotkeys.h"
 #include "game_interface.h"
 #include "icn.h"
 #include "localevent.h"
