+++ src/fheroes2/dialog/dialog_thievesguild.cpp
@@ -29,6 +29,7 @@
 #include "cursor.h"
 #include "dialog.h"
 #include "game.h"
+#include "game_hotkeys.h"
 #include "icn.h"
 #include "kingdom.h"
 #include "monster.h"
