+++ src/fheroes2/game/game_highscores.cpp
@@ -33,6 +33,7 @@
 #include "dialog.h"
 #include "game.h"
 #include "game_delays.h"
+#include "game_hotkeys.h"
 #include "game_over.h"
 #include "icn.h"
 #ifdef WITH_DEBUG
