+++ src/fheroes2/game/game_interface.cpp
@@ -27,6 +27,7 @@
 #include "dialog.h"
 #include "game.h"
 #include "game_delays.h"
+#include "game_hotkeys.h"
 #include "icn.h"
 #include "maps.h"
 #include "settings.h"
