+++ src/fheroes2/game/game_loadgame.cpp
@@ -27,6 +27,7 @@
 #include "audio.h"
 #include "cursor.h"
 #include "dialog.h"
+#include "game_hotkeys.h"
 #include "game_io.h"
 #include "game_mainmenu_ui.h"
 #include "icn.h"
