+++ src/fheroes2/gui/ui_button.cpp
@@ -21,7 +21,7 @@
 #include "ui_button.h"
 #include "agg_image.h"
 #include "dialog.h"
-#include "game.h"
+#include "game_hotkeys.h"
 #include "icn.h"
 #include "localevent.h"
 #include "pal.h"
