+++ src/fheroes2/gui/ui_dialog.cpp
@@ -23,8 +23,8 @@
 #include "cursor.h"
 #include "dialog.h"
 #include "experience.h"
-#include "game.h"
 #include "game_delays.h"
+#include "game_hotkeys.h"
 #include "heroes_indicator.h"
 #include "icn.h"
 #include "localevent.h"
