+++ src/fheroes2/heroes/heroes_dialog.cpp
@@ -27,7 +27,7 @@
 #include "army_bar.h"
 #include "cursor.h"
 #include "dialog.h"
-#include "game.h"
+#include "game_hotkeys.h"
 #include "heroes.h"
 #include "heroes_indicator.h"
 #include "icn.h"
