+++ src/fheroes2/kingdom/puzzle.cpp
@@ -29,6 +29,7 @@
 #include "cursor.h"
 #include "game.h"
 #include "game_delays.h"
+#include "game_hotkeys.h"
 #include "game_interface.h"
 #include "icn.h"
 #include "logging.h"
